+++ src/builder.rs
@@ -62,16 +62,19 @@ impl Builder {
         T: Send + 'static,
     {
         let (join_check_send, join_check_receive) = ::crossbeam_channel::bounded(1);
+        let full_name = self.full_name;
+        let name = self.name;
         let shutdown = Arc::new(AtomicBool::new(false));
         let scope_shutdown = Arc::clone(&shutdown);
-        let status = RegisteredStatus::new(self.name, self.full_name);
         let join = self
             .std
             .spawn(|| {
                 let id = current_thread_id();
+                let status = RegisteredStatus::new(name, full_name);
+                let activity = status.activity();
                 // Keep a ThreadGuard alive as long as the thread is.
                 let _guard = ThreadGuard::new(id, join_check_send, status);
-                let scope = ThreadScope::new(scope_shutdown);
+                let scope = ThreadScope::new(activity, scope_shutdown);
                 f(scope)
             })
             .with_context(|_| ErrorKind::Spawn)?;