+++ src/lib.rs
@@ -1,3 +1,5 @@
+#![doc(html_root_url = "https://docs.rs/humthreads/0.1.0")]
+
 extern crate crossbeam_channel;
 extern crate failure;
 extern crate lazy_static;
@@ -17,5 +19,6 @@ pub use self::error::ErrorKind;
 pub use self::error::Result;
 pub use self::handles::Thread;
 pub use self::handles::ThreadScope;
+pub use self::handles::ThreadScopeActivityGuard;
 pub use self::registry::registered_threads;
 pub use self::status::ThreadStatus;