+++ src/status.rs
@@ -1,18 +1,37 @@
+use std::sync::Arc;
+use std::sync::Mutex;
+
 /// Internal status traking for registered threads.
 pub(crate) struct RegisteredStatus {
+    activity: Arc<Mutex<Option<String>>>,
     name: String,
     short_name: String,
 }
 
 impl RegisteredStatus {
+    /// Provide mutable access to the thread's activity status attribute.
+    pub(crate) fn activity(&self) -> Arc<Mutex<Option<String>>> {
+        Arc::clone(&self.activity)
+    }
+
     pub(crate) fn new(name: String, short_name: String) -> RegisteredStatus {
-        RegisteredStatus { name, short_name }
+        let activity = Arc::new(Mutex::new(None));
+        RegisteredStatus {
+            activity,
+            name,
+            short_name,
+        }
     }
 }
 
 /// Public view of a point in time status of a thread.
 #[derive(Clone, Eq, PartialEq, Hash, Debug, Serialize, Deserialize)]
 pub struct ThreadStatus {
+    /// Description of the activity currently in progress by the thread.
+    ///
+    /// NOTE: threads are responsable for reporting their own activity.
+    pub activity: Option<String>,
+
     /// Full name of the thread.
     pub name: String,
 
@@ -24,9 +43,38 @@ pub struct ThreadStatus {
 
 impl From<&RegisteredStatus> for ThreadStatus {
     fn from(status: &RegisteredStatus) -> ThreadStatus {
+        let activity = status
+            .activity
+            .lock()
+            .expect("RegisteredStatus::activity lock poisoned")
+            .clone();
         ThreadStatus {
+            activity,
             name: status.name.clone(),
             short_name: status.short_name.clone(),
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::RegisteredStatus;
+    use super::ThreadStatus;
+
+    #[test]
+    fn from_register() {
+        let register = RegisteredStatus::new("long name".into(), "name".into());
+        let status = ThreadStatus::from(&register);
+        assert_eq!(status.activity, None);
+        assert_eq!(status.name, "long name");
+        assert_eq!(status.short_name, "name");
+    }
+
+    #[test]
+    fn report_activity() {
+        let register = RegisteredStatus::new("long name".into(), "name".into());
+        *register.activity.lock().unwrap() = Some("test".into());
+        let status = ThreadStatus::from(&register);
+        assert_eq!(status.activity, Some("test".into()));
     }
 }